Move more code into buildSrc
