+++ app/build.gradle.kts
@@ -1,5 +1,3 @@
-import org.apache.tools.ant.filters.FixCrLfFilter
-
 plugins {
     id("com.android.application")
     kotlin("android")
@@ -19,8 +17,6 @@ kapt {
         arg("room.incremental", "true")
     }
 }
-
-setupApp()
 
 android {
     defaultConfig {
@@ -68,93 +64,7 @@ android {
     }
 }
 
-val syncLibs by tasks.registering(Sync::class) {
-    into("src/main/jniLibs")
-    into("armeabi-v7a") {
-        from(rootProject.file("native/out/armeabi-v7a")) {
-            include("busybox", "magiskboot", "magiskinit", "magisk")
-            rename { if (it == "magisk") "libmagisk32.so" else "lib$it.so" }
-        }
-    }
-    into("x86") {
-        from(rootProject.file("native/out/x86")) {
-            include("busybox", "magiskboot", "magiskinit", "magisk")
-            rename { if (it == "magisk") "libmagisk32.so" else "lib$it.so" }
-        }
-    }
-    into("arm64-v8a") {
-        from(rootProject.file("native/out/arm64-v8a")) {
-            include("busybox", "magiskboot", "magiskinit", "magisk")
-            rename { if (it == "magisk") "libmagisk64.so" else "lib$it.so" }
-        }
-    }
-    into("x86_64") {
-        from(rootProject.file("native/out/x86_64")) {
-            include("busybox", "magiskboot", "magiskinit", "magisk")
-            rename { if (it == "magisk") "libmagisk64.so" else "lib$it.so" }
-        }
-    }
-    onlyIf {
-        if (inputs.sourceFiles.files.size != 16)
-            throw StopExecutionException("Please build binaries first! (./build.py binary)")
-        true
-    }
-}
-
-val syncAssets by tasks.registering(Sync::class) {
-    dependsOn(syncLibs)
-    inputs.property("version", Config.version)
-    inputs.property("versionCode", Config.versionCode)
-    into("src/main/assets")
-    from(rootProject.file("scripts")) {
-        include("util_functions.sh", "boot_patch.sh", "uninstaller.sh", "addon.d.sh")
-    }
-    into("chromeos") {
-        from(rootProject.file("tools/futility"))
-        from(rootProject.file("tools/keys")) {
-            include("kernel_data_key.vbprivk", "kernel.keyblock")
-        }
-    }
-    filesMatching("**/util_functions.sh") {
-        filter {
-            it.replace(
-                "#MAGISK_VERSION_STUB",
-                "MAGISK_VER='${Config.version}'\nMAGISK_VER_CODE=${Config.versionCode}"
-            )
-        }
-        filter<FixCrLfFilter>("eol" to FixCrLfFilter.CrLf.newInstance("lf"))
-    }
-}
-
-val syncResources by tasks.registering(Sync::class) {
-    dependsOn(syncAssets)
-    into("src/main/resources/META-INF/com/google/android")
-    from(rootProject.file("scripts/update_binary.sh")) {
-        rename { "update-binary" }
-    }
-    from(rootProject.file("scripts/flash_script.sh")) {
-        rename { "updater-script" }
-    }
-}
-
-tasks.preBuild {
-    dependsOn(syncResources)
-}
-
-android.applicationVariants.all {
-    val keysDir = rootProject.file("tools/keys")
-    val outSrcDir = File(buildDir, "generated/source/keydata/$name")
-    val outSrc = File(outSrcDir, "com/topjohnwu/magisk/signing/KeyData.java")
-
-    val genSrcTask = tasks.register("generate${name.capitalize()}KeyData") {
-        inputs.dir(keysDir)
-        outputs.file(outSrc)
-        doLast {
-            genKeyData(keysDir, outSrc)
-        }
-    }
-    registerJavaGeneratingTask(genSrcTask, outSrcDir)
-}
+setupApp()
 
 configurations.all {
     exclude("org.jetbrains.kotlin", "kotlin-stdlib-jdk7")
@@ -207,7 +117,7 @@ dependencies {
     implementation("androidx.room:room-ktx:${vRoom}")
     kapt("androidx.room:room-compiler:${vRoom}")
 
-    val vNav = "2.4.0-alpha07"
+    val vNav = "2.4.0-alpha08"
     implementation("androidx.navigation:navigation-fragment-ktx:${vNav}")
     implementation("androidx.navigation:navigation-ui-ktx:${vNav}")
 
@@ -218,7 +128,7 @@ dependencies {
     implementation("androidx.preference:preference:1.1.1")
     implementation("androidx.recyclerview:recyclerview:1.2.1")
     implementation("androidx.fragment:fragment-ktx:1.3.6")
-    implementation("androidx.work:work-runtime-ktx:2.7.0-alpha05")
+    implementation("androidx.work:work-runtime-ktx:2.7.0-beta01")
     implementation("androidx.transition:transition:1.4.1")
     implementation("androidx.core:core-ktx:1.6.0")
     implementation("com.google.android.material:material:1.4.0")
