+++ buildSrc/build.gradle.kts
@@ -18,7 +18,7 @@ gradlePlugin {
 dependencies {
     implementation(kotlin("gradle-plugin", "1.5.30"))
     implementation("com.android.tools.build:gradle:7.0.2")
-    implementation("androidx.navigation:navigation-safe-args-gradle-plugin:2.4.0-alpha07")
+    implementation("androidx.navigation:navigation-safe-args-gradle-plugin:2.4.0-alpha08")
     implementation("io.michaelrocks:paranoid-gradle-plugin:0.3.5")
     implementation("org.eclipse.jgit:org.eclipse.jgit:5.12.0.202106070339-r")
 }
